refactor(skia): move getSquirclePath to worklets/skia
Move getSquirclePath from SkiaCard.tsx to src/worklets/skia.ts where it belongs alongside other Skia path utilities like getCirclePath.

This function is a generic geometric utility with no card-specific logic, and was only in `SkiaCard.tsx` because that's where it was first written. The new location makes the architecture clearer and removes unnecessary dependencies on the unused skia-cards components, which we can then remove in isolation in an upcoming PR.
